Правка после проверки 1
1. Модули get-random-positive-float и get-random-positive-integer убрала, а функции поместила в utils.

2. Константы и функции, относящиеся к логике объявлений, перенесла из модуля utils в модуль data.
